+++ testing/constants.py
@@ -18,3 +18,12 @@ unmap_addr_fmt = {
     'p2wsh': AF_P2WSH,
     'p2wsh-p2sh': AF_P2WSH_P2SH,
 }
+
+# all possible addr types, including multisig/scripts
+ADDR_STYLES = ['p2wpkh', 'p2wsh', 'p2sh', 'p2pkh', 'p2wsh-p2sh', 'p2wpkh-p2sh']
+
+# single-signer
+ADDR_STYLES_SINGLE = ['p2wpkh', 'p2pkh', 'p2wpkh-p2sh']
+
+# multi signer
+ADDR_STYLES_MS = ['p2sh', 'p2wsh', 'p2wsh-p2sh']